requirements.txt
numpy
rdkit
mordredcommunity
torch
deepchem
scikit-learn
mhfp
<<<<<<< HEAD
pandas
=======
pandas
>>>>>>> 9fb59ae (Local changes and new files)
